Update README.md
My group members and myself have done some digging and noticed there are a few other healthcare related ideas that have been done in the past. A software that allows facilities to keep up with their budget would be a valuable tool. This could monitor the incoming revenue of the facility, the necessary expensive to ensure the facility is up-to-date, employee budgets, etc.
A clinic will be able to identify what time of the year brings in the most revenue which would mainly be towards the end of the year due to insurance liabilities being met. This will also keep the office updated in how much was lost from any mishaps and give them an idea of what can be scheduled to obtain the loss.